requirements.txt
# requirements.txt
openai
python-dotenv
google-generativeai